Tidy release tags and separate update code
Strips out any leading 'v' or "data release" text from the release tags
to prevent minor version tag formatting issues such as #296

Also moves update code to a separate utils and tidies the resulting
imports in command.py
